us#171024 replace avatars with library components
In us#171024, we replace the avatars with the Avatar component from the library throughout the Targetprocess code. This PR makes the avatars generated with Custom Unit All Assigned People mashup consistent with other TP units with avatars.

This update should come to the user with the same TP update that will ship the us#171024 changes.