Update ZoomButtons.js
Refactored code, cleaner. Moved CSS Selectors to top of script for easier maintainance. Added more comments to make it easier to understand what the script is doing. Timeouts and pollig is minimal and cannot be avoided without direct API access to the zoom features. Also the Menu is only in the DOM when the action-button is clicked and then removed. The slider form is temporary.